requirements.txt
# Smart CCTV System - Requirements

# Core Dependencies
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLO v8
loguru>=0.7.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Audio & Alerts
pyttsx3>=2.90
PyAudio>=0.2.13
pydub>=0.25.1

# Tracking
scipy>=1.10.0
scikit-learn>=1.3.0
filterpy>=1.4.5

# Data & Storage
pandas>=2.0.0
Pillow>=10.0.0

# Utilities
imutils>=0.5.4
psutil>=5.9.0
requests>=2.31.0

# Web Interface (Optional)
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0

# Monitoring (Optional)
prometheus-client>=0.17.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0

# Platform-specific
# Linux only:
# PyBluez>=0.23  # Bluetooth support

# Windows only:
# pywin32>=306  # Windows API

# macOS:
# pyobjc>=9.2  # macOS APIs
